MLDL/ML_PROJECT/spam-filter/requirements.txt
flask
pandas
scikit-learn
nltk
joblib
python-dotenv
flask-sqlalchemy
flask-login
flask-bcrypt
